pyproject.toml
[project]
name = "100-days-of-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.12"
dependencies = [
    "ruff",
    "pylint",
    "colorgram.py",
    "heroes",
    "villains",
    "pandas",
    "prettytable",
    "pyperclip"
]
